requirements.txt
beautifulsoup4==4.14.2
lxml==6.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
